+++ src/main/kotlin/app/revanced/cli/main/Main.kt
@@ -1,8 +1,0 @@
-package app.revanced.cli.main
-
-import app.revanced.cli.command.MainCommand
-import picocli.CommandLine
-
-internal fun main(args: Array<String>) {
-    CommandLine(MainCommand).execute(*args)
-}