+++ src/main/kotlin/app/revanced/utils/adb/Commands.kt
@@ -5,10 +5,9 @@ import se.vidstige.jadb.RemoteFile
 import se.vidstige.jadb.ShellProcessBuilder
 import java.io.File
 
+
 internal fun JadbDevice.buildCommand(command: String, su: Boolean = true): ShellProcessBuilder {
-    if (su) {
-        return shellProcessBuilder("su -c \'$command\'")
-    }
+    if (su) return shellProcessBuilder("su -c \'$command\'")
 
     val args = command.split(" ") as ArrayList<String>
     val cmd = args.removeFirst()
@@ -20,10 +19,15 @@ internal fun JadbDevice.run(command: String, su: Boolean = true): Int {
     return this.buildCommand(command, su).start().waitFor()
 }
 
-internal fun JadbDevice.copy(targetPath: String, file: File) {
-    push(file, RemoteFile(targetPath))
-}
+internal fun JadbDevice.hasSu() =
+    this.startCommand("su -h", false).waitFor() == 0
 
-internal fun JadbDevice.createFile(targetFile: String, content: String) {
+internal fun JadbDevice.push(file: File, targetFilePath: String) =
+    push(file, RemoteFile(targetFilePath))
+
+internal fun JadbDevice.createFile(targetFile: String, content: String) =
     push(content.byteInputStream(), System.currentTimeMillis(), 644, RemoteFile(targetFile))
-}
+
+
+private fun JadbDevice.startCommand(command: String, su: Boolean) =
+    shellProcessBuilder(if (su) "su -c '$command'" else command).start()