+++ src/main/kotlin/app/revanced/utils/align/zip/structures/ZipEndRecord.kt
@@ -1,9 +1,9 @@
-package app.revanced.utils.signing.align.zip.structures
+package app.revanced.utils.align.zip.structures
 
-import app.revanced.utils.signing.align.zip.putUInt
-import app.revanced.utils.signing.align.zip.putUShort
-import app.revanced.utils.signing.align.zip.readUIntLE
-import app.revanced.utils.signing.align.zip.readUShortLE
+import app.revanced.utils.align.zip.putUInt
+import app.revanced.utils.align.zip.putUShort
+import app.revanced.utils.align.zip.readUIntLE
+import app.revanced.utils.align.zip.readUShortLE
 import java.io.DataInput
 import java.nio.ByteBuffer
 import java.nio.ByteOrder