+++ src/main/kotlin/app/revanced/utils/signing/ApkSigner.kt
@@ -1,7 +1,5 @@
 package app.revanced.utils.signing
 
-import app.revanced.cli.command.MainCommand.logger
-import app.revanced.cli.signing.SigningOptions
 import com.android.apksig.ApkSigner
 import org.bouncycastle.asn1.x500.X500Name
 import org.bouncycastle.asn1.x509.SubjectPublicKeyInfo
@@ -17,11 +15,44 @@ import java.math.BigInteger
 import java.security.*
 import java.security.cert.X509Certificate
 import java.util.*
+import java.util.logging.Logger
 
-internal class Signer(
+internal class ApkSigner(
     private val signingOptions: SigningOptions
 ) {
+    private val logger = Logger.getLogger(ApkSigner::class.java.name)
+
+    private val signer: ApkSigner.Builder
     private val passwordCharArray = signingOptions.password.toCharArray()
+
+    init {
+        Security.addProvider(BouncyCastleProvider())
+
+        val keyStore = KeyStore.getInstance("BKS", "BC")
+        val alias = keyStore.let { store ->
+            FileInputStream(File(signingOptions.keyStoreFilePath).also {
+                if (!it.exists()) {
+                    logger.info("Creating keystore at ${it.absolutePath}")
+                    newKeystore(it)
+                } else {
+                    logger.info("Using keystore at ${it.absolutePath}")
+                }
+            }).use { fis -> store.load(fis, null) }
+            store.aliases().nextElement()
+        }
+
+        with(
+            ApkSigner.SignerConfig.Builder(
+                signingOptions.cn,
+                keyStore.getKey(alias, passwordCharArray) as PrivateKey,
+                listOf(keyStore.getCertificate(alias) as X509Certificate)
+            ).build()
+        ) {
+            this@ApkSigner.signer = ApkSigner.Builder(listOf(this))
+            signer.setCreatedBy(signingOptions.cn)
+        }
+    }
+
     private fun newKeystore(out: File) {
         val (publicKey, privateKey) = createKey()
         val privateKS = KeyStore.getInstance("BKS", "BC")
@@ -50,30 +81,12 @@ internal class Signer(
         return JcaX509CertificateConverter().getCertificate(builder.build(signer)) to pair.private
     }
 
-    fun signApk(input: File, output: File) {
-        Security.addProvider(BouncyCastleProvider())
-
-        // TODO: keystore should be saved securely
-        val ks = File(signingOptions.keyStoreFilePath)
-        if (!ks.exists()) newKeystore(ks) else {
-            logger.info("Found existing keystore: ${ks.name}")
-        }
-
-        val keyStore = KeyStore.getInstance("BKS", "BC")
-        FileInputStream(ks).use { fis -> keyStore.load(fis, null) }
-        val alias = keyStore.aliases().nextElement()
-
-        val config = ApkSigner.SignerConfig.Builder(
-            signingOptions.cn,
-            keyStore.getKey(alias, passwordCharArray) as PrivateKey,
-            listOf(keyStore.getCertificate(alias) as X509Certificate)
-        ).build()
-
-        val signer = ApkSigner.Builder(listOf(config))
-        signer.setCreatedBy(signingOptions.cn)
+    fun signApk(input: File, output: File): File {
         signer.setInputApk(input)
         signer.setOutputApk(output)
 
         signer.build().sign()
+
+        return output
     }
 }